+++ apps/core-lib-dev/src/app/datepicker.element.ts
@@ -2,7 +2,9 @@ import { LitElement } from 'lit'
 import { customElement, state } from 'lit/decorators.js'
 import { html } from '@sebgroup/green-core/scoping'
 
-import { registerTransitionalStyles } from '@sebgroup/green-core/transitional-styles'
+import '@sebgroup/green-core/components/datepicker/index.js'
+import * as DatepickerTransitionalStyles from '@sebgroup/green-core/components/datepicker/datepicker.trans.styles.js'
+DatepickerTransitionalStyles.register()
 
 @customElement('datepicker-example')
 export class CalendarExample extends LitElement {
@@ -12,7 +14,6 @@ export class CalendarExample extends LitElement {
 
   connectedCallback() {
     super.connectedCallback()
-    registerTransitionalStyles()
   }
 
   @state()