+++ libs/angular/src/lib/in-page-wizard/in-page-wizard-step-card.spec.ts
@@ -73,7 +73,7 @@ describe('InPageWizardStepCardComponent', () => {
       })
     })
 
-    it('should show next button', async () => {
+    it('should show next button with icon', async () => {
       const { findByTestId } = await render(
         NggFakeInPageWizardStepCardComponent,
         {
@@ -92,9 +92,11 @@ describe('InPageWizardStepCardComponent', () => {
       const nextBtn = await findByTestId('in-page-wizard-step-card-next-btn')
       expect(nextBtn).toBeDefined()
       expect(nextBtn.textContent?.trim()).toEqual('Test Next')
+      const nextIcon = await findByTestId('in-page-wizard-step-card-next-icon')
+      expect(nextIcon).toBeDefined()
     })
 
-    it('should not show edit btn', async () => {
+    it('should not show any edit btns', async () => {
       const { queryByTestId } = await render(
         NggFakeInPageWizardStepCardComponent,
         {
@@ -112,6 +114,8 @@ describe('InPageWizardStepCardComponent', () => {
 
       const el = queryByTestId('in-page-wizard-step-card-edit-btn')
       expect(el).toBeNull()
+      const elFooter = queryByTestId('in-page-wizard-step-card-footer-edit-btn')
+      expect(elFooter).toBeNull()
     })
 
     it('should show content', async () => {
@@ -389,6 +393,8 @@ class NggFakeInPageWizardStepCardComponent {
   @Input() public editBtnText = ''
 
   @Input() public nextBtnText = ''
+  
+  @Input() public showNextBtnIcon = true
 
   @Input() public isCompleted = false