+++ libs/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.5.0
+
+### Minor Changes
+
+- 70bbc39: New components Card, Container, IMG, Video and tokens upgrade
+
+## 0.4.1
+
+### Patch Changes
+
+- 4a4a67c: Fixed 2016 base scale dark mode colors
+
 ## 0.4.0
 
 ### Minor Changes